New features: add customer, group customers; create, delete and edit group, create preset, delete template

+++ src/app/lib/actions.ts
@@ -68,12 +68,256 @@ const CreateTemplateSchema = z.object({
     ),
 });
 
+const AddCustomerSchema = z.object({
+  name: z
+    .string()
+    .min(4, { message: "Name must be at least 4 characters long" }),
+  email: z
+    .string()
+    .min(1, { message: "E-mail is required" })
+    .email({ message: "Invalid e-mail" }),
+  company: z.string().optional(),
+});
+
 const SendNewsletterPayloadSchema = z.object({
   templateId: z.string().min(1, { message: "Please select a template" }),
   customers: z.array(z.string()).nonempty({
     message: "Please select a customer.",
   }),
 });
+const GroupCustomersPayloadSchema = z.object({
+  groupId: z.string().min(1, { message: "Please select a group" }),
+  customers: z.array(z.string()).nonempty({
+    message: "Please select a customer.",
+  }),
+});
+
+export async function createPreset(
+  templateId: string,
+  templateFormState: any,
+  _currentState: unknown,
+  formData: FormData
+) {
+  const session = await getSession();
+  const userId = session?.user?.id || "";
+
+  let template = null;
+  const shape: any = {};
+
+  if (!templateId) {
+    return {
+      message: "Please select a template",
+      success: false,
+    };
+  }
+
+  const presetNameSchema = z.object({
+    presetName: z.string().min(4, {
+      message: "Invalid Name. Name must be at least 4 characters long",
+    }),
+  });
+  const validatedName = presetNameSchema.safeParse({
+    presetName: formData.get("presetName"),
+  });
+  if (!validatedName.success) {
+    console.log(validatedName.error.flatten().fieldErrors);
+    return {
+      errors: validatedName.error.flatten().fieldErrors,
+      message: "Missing Fields. Failed to create a preset.",
+      success: false,
+    };
+  }
+
+  const { presetName } = validatedName.data;
+
+  try {
+    const results: any = await query(
+      "SELECT name FROM presets WHERE name=? AND user_id=?",
+      [presetName, userId]
+    );
+
+    if (results.length > 0) {
+      return {
+        message: "Preset with that name already exists",
+        success: false,
+      };
+    }
+  } catch (error) {
+    console.log(error);
+    return {
+      message: "Database Error. Failed to fetch preset",
+      success: false,
+    };
+  }
+
+  try {
+    template = await fetchTemplateById(templateId);
+  } catch (error) {
+    console.log(error);
+    return {
+      message: "Database Error. Failed to fetch template",
+      success: false,
+    };
+  }
+  const { inputs, html } = template;
+
+  for (const input of inputs) {
+    if (input.originalTag === "a") {
+      shape[input.name] = z.string().url({ message: "Invalid url format" });
+    } else {
+      shape[input.name] = z
+        .string()
+        .min(5, { message: "Please provide at least 5 letters" });
+    }
+  }
+
+  // validate template form state
+  const TemplateFormSchema = z.object(shape);
+  const validatedTemplateFormState =
+    TemplateFormSchema.safeParse(templateFormState);
+
+  if (!validatedTemplateFormState.success) {
+    console.log(validatedTemplateFormState.error.flatten().fieldErrors);
+    return {
+      errors: validatedTemplateFormState.error.flatten().fieldErrors,
+      message: "Missing Fields. Failed to create a preset.",
+      success: false,
+    };
+  }
+
+  const { data } = validatedTemplateFormState;
+
+  // sanitize template form state data
+  const sanitizedData: any = {};
+  for (const field in data) {
+    if (Object.prototype.hasOwnProperty.call(data, field)) {
+      const value = data[field];
+      sanitizedData[field] = DOMPurify.sanitize(value);
+    }
+  }
+  const readyToSaveHTML = injectDataIntoHTML(data, html);
+
+  try {
+    await query("INSERT INTO presets (user_id, html, name) VALUES(?, ?, ?)", [
+      userId,
+      readyToSaveHTML,
+      presetName,
+    ]);
+  } catch (error) {
+    console.log(error);
+    return {
+      message: "Database Error. Failed to create a preset",
+      success: false,
+    };
+  }
+
+  revalidatePath("/dashboard/send-newsletter");
+  return { message: "Preset created successfully", success: true };
+}
+
+export async function addCustomer(_currentState: unknown, formData: FormData) {
+  const validatedFormData = AddCustomerSchema.safeParse({
+    name: formData.get("name") || "",
+    email: formData.get("email") || "",
+    company: formData.get("company") || "",
+  });
+
+  if (!validatedFormData.success) {
+    console.log(validatedFormData.error.flatten().fieldErrors);
+    return {
+      errors: validatedFormData.error.flatten().fieldErrors,
+      message: "Invalid Fields. Failed to create a customer",
+      success: false,
+    };
+  }
+
+  const { name, email, company } = validatedFormData.data;
+
+  try {
+    const results: any = await query(
+      "SELECT email FROM customers WHERE email=?",
+      [email]
+    );
+    if (results.length > 0) {
+      return {
+        message: "A customer with given e-mail already exists.",
+        success: false,
+      };
+    }
+  } catch (error) {
+    console.log(error);
+    return {
+      message: "Database Error: Failed to return a customer",
+      success: false,
+    };
+  }
+
+  try {
+    await query(
+      "INSERT INTO customers (name, email, company) VALUES(?, ?, ?)",
+      [name, email, company]
+    );
+  } catch (error) {
+    console.log(error);
+    return {
+      message: "Database Error. Failed to create a customer",
+      success: false,
+    };
+  }
+
+  revalidatePath("/dashboard/send-newsletter");
+  return {
+    message: "Customer successfully created",
+    success: true,
+  };
+}
+
+export async function groupCustomers(payload: any) {
+  const validatedPayload = GroupCustomersPayloadSchema.safeParse(payload);
+  if (!validatedPayload.success) {
+    console.log(validatedPayload.error.flatten().fieldErrors);
+    return {
+      errors: validatedPayload.error.flatten().fieldErrors,
+      message: "Invalid Fields. Failed to Group Customers.",
+      success: false,
+    };
+  }
+
+  const { groupId, customers } = validatedPayload.data;
+
+  try {
+    for (const customerId of customers) {
+      const results: any = await query(
+        "SELECT * FROM grouped_customers WHERE group_id=? AND customer_id=?",
+        [groupId, customerId]
+      );
+
+      if (results.length < 1) {
+        await query(
+          "INSERT INTO grouped_customers (group_id, customer_id) VALUES (?, ?) ",
+          [groupId, customerId]
+        );
+      } else {
+        return {
+          message: "Customer already exists in that group",
+          success: false,
+        };
+      }
+    }
+  } catch (error) {
+    console.log(error);
+    return {
+      message: "Database Error: Failed to save customers to a group",
+      success: false,
+    };
+  }
+
+  revalidatePath("/dashboard/send-newsletter");
+  return {
+    message: "Customers successfully saved to group",
+    success: true,
+  };
+}
 
 export async function createGroup(_currentState: unknown, formData: FormData) {
   const session = await getSession();
@@ -120,7 +364,7 @@ export async function createGroup(_currentState: unknown, formData: FormData) {
     };
   }
 
-  revalidatePath("/send-newsletter");
+  revalidatePath("/dashboard/send-newsletter");
   return { message: "Group created successfully", success: true };
 }
 
@@ -181,13 +425,20 @@ export async function editGroup(
     };
   }
 
-  revalidatePath("/send-newsletter");
+  revalidatePath("/dashboard/send-newsletter");
   return { message: "Group edited successfully", success: true };
 }
 
 export async function deleteGroup(groupId: string, _currentState: unknown) {
   const session = await getSession();
   const userId = session?.user?.id || "";
+
+  if (!groupId) {
+    return {
+      message: "Invalid group id. Failed to delete group",
+      success: false,
+    };
+  }
 
   try {
     await query("DELETE FROM grouped_customers WHERE group_id=?", [groupId]);
@@ -203,8 +454,49 @@ export async function deleteGroup(groupId: string, _currentState: unknown) {
     };
   }
 
-  revalidatePath("/send-newsletter");
+  revalidatePath("/dashboard/send-newsletter");
   return { message: "Group deleted successfully", success: true };
+}
+
+export async function deleteTemplate(templateId: string) {
+  const session = await getSession();
+  const userId = session?.user?.id || "";
+
+  if (!templateId) {
+    return { message: "Invalid template id. Please try again", success: false };
+  }
+
+  try {
+    const result: any = await query(
+      "SELECT id FROM templates WHERE id=? AND user_id=?",
+      [templateId, userId]
+    );
+    if (result.length < 1) {
+      return { message: "Template does not exist", success: false };
+    }
+  } catch (error) {
+    console.log(error);
+    return {
+      message: "Database Error. Failed to fetch template",
+      success: false,
+    };
+  }
+
+  try {
+    await query("DELETE FROM templates where id=? AND user_id=?", [
+      templateId,
+      userId,
+    ]);
+  } catch (error) {
+    return {
+      message: "Database Error. Failed to delete template",
+      success: false,
+    };
+  }
+
+  revalidatePath("/dashboard/templates");
+  revalidatePath("/dashboard/send-newsletter");
+  return { message: "Template deleted sucessfully", success: true };
 }
 
 export async function createTemplate(
